requirements.txt
google.generativeai
matplotlib
pandas
streamlit
plotly
numpy
Pillow
openpyxl
streamlit_sortables
networkx
workadays
streamlit_extras
tabulate